move hook to hooks
